services/web/requirements.txt
boto3==1.23.10
Flask==2.0.3
marshmallow==3.14.1
PyJWT==2.4.0
cryptography==37.0.4
gunicorn==20.1.0
Flask-Cors==3.0.10